Renamed TorchDotNet into TorchTensor.
